LayoutTests/inspector/debugger/break-on-uncaught-exception-expected.txt

CONSOLE MESSAGE: throwing TestError
CONSOLE MESSAGE: TestError
CONSOLE MESSAGE: testFinally
CONSOLE MESSAGE: throwing TestError
CONSOLE MESSAGE: finally
CONSOLE MESSAGE: TestError
CONSOLE MESSAGE: testThrowingThruNativeCode
CONSOLE MESSAGE: throwing TestError
CONSOLE MESSAGE: TestError
CONSOLE MESSAGE: testCatch
CONSOLE MESSAGE: throwing TestError
CONSOLE MESSAGE: catch TestError
CONSOLE MESSAGE: DONE
Checking pause locations when pausing on uncaught exceptions.


== Running test suite: BreakOnUncaughtException
-- Running test case: BreakOnUncaughtException.UncaughtExceptionBasic
PASS: Should pause for exception.
PAUSE AT doThrow:8:22
Uncaught exception in test page: TestError [break-on-exception-tests.js:8]

-- Running test case: BreakOnUncaughtException.UncaughtExceptionInFinally
PASS: Should pause for exception.
PAUSE AT doThrow:8:22
Uncaught exception in test page: TestError [break-on-exception-tests.js:8]

-- Running test case: BreakOnUncaughtException.UncaughtExceptionThruNativeCode
PASS: Should pause for exception.
PAUSE AT doThrow:8:22
Uncaught exception in test page: TestError [break-on-exception-tests.js:8]

-- Running test case: BreakOnUncaughtException.CaughtException
PASS: Should not pause, exception was caught.
